+++ src/pages/Home.jsx
@@ -15,3 +15,4 @@ function Home() {
 }
 
 export default Home;
+