+++ CHANGELOG.md
@@ -1,5 +1,88 @@
 # Changelog
 
+## [1.10.1](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/compare/v1.10.0...v1.10.1) (2024-07-21)
+
+
+### Bug Fixes
+
+* **inventory:** Equipment changing fix. ([9b9d1d6](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/9b9d1d6dcb12d62bcada43b2de0de4bd0c8667ce))
+
+## [1.10.0](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/compare/v1.9.1...v1.10.0) (2024-07-20)
+
+
+### Features
+
+* **Game:** Right click context menu. ([cf872ad](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/cf872ade4f95b0a0cba7675f1b72ae8b45f2c26b))
+* **Inventory:** Open menu context. ([02ffe42](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/02ffe42c49e1ca3ced1e358df031b93839f022fe))
+* **Inventory:** Sell menu ([2a50018](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/2a50018c0adcd8bf70054fa6b9eaa107744e053a))
+
+## [1.9.1](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/compare/v1.9.0...v1.9.1) (2024-07-20)
+
+
+### Bug Fixes
+
+* **Expeditions:** Progress shows current level ([2e27f57](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/2e27f575ec3384be121e57e9f55a5d70ef14a7c1))
+
+## [1.9.0](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/compare/v1.8.1...v1.9.0) (2024-07-20)
+
+
+### Features
+
+* **expeditions:** Drop rate tooltip. ([111c1b2](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/111c1b276c437b3456f1f78b81e768e6700eb2da))
+* **Expeditions:** Level locked areas. ([0ed587b](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/0ed587b8c95f287b043ef6832b8696b0027cc8db))
+
+
+### Bug Fixes
+
+* **leveling:** Off by one. ([fc40a64](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/fc40a641c164d89ff04a1b086036054b686949dc))
+
+## [1.8.1](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/compare/v1.8.0...v1.8.1) (2024-07-20)
+
+
+### Bug Fixes
+
+* **inventory:** display players name properly. ([40a3848](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/40a3848d240c369a9a0a8f138bcd378c5d076c99))
+
+## [1.8.0](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/compare/v1.7.3...v1.8.0) (2024-07-16)
+
+
+### Features
+
+* **main:** version display. ([b32dcbd](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/b32dcbd01eb3b964eb67afb2dd41480766eea248))
+* **workflow:** auto pr merger for version. ([3cf21aa](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/3cf21aa72caee60f27ae947471d1f8f1ab4f712a))
+
+
+### Bug Fixes
+
+* **auto-merge.yml:** test ([ffe1433](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/ffe1433946ed55aafe3e3d755c292a1c0159f86a))
+* **automerge:** test 123 ([aec1a07](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/aec1a077d1bad5b5f37bca25782d0dd01558298a))
+* **autoPR:** hopefully ([d19639e](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/d19639e327c9ea44b14197e4f87fdd9b3a8f86f5))
+* **Expeditions:** Adds progress bar. ([f89231f](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/f89231fb996df0fe57283c574537ad6cd35ebbee))
+* **expeditions:** again ([ecf0e98](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/ecf0e98fd885d6e94d951407298177c590aa3f12))
+* **expeditions:** double reward fix ([2d5e83f](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/2d5e83ffa5443970d1a915c244bfefbeee27cd94))
+* **expeditions:** Expedition start time fix. ([b4d6a89](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/b4d6a89cc441c719fa4c29991d7323ca6e216d5a))
+* **expeditions:** interval didn't properly stop ([83e1d14](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/83e1d149dcca4532177f5e953f708ce42867de92))
+* **Expeditions:** progress bar reset when leaving ([c9346be](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/c9346be948149edb3bb9f40ee80c7e114e069edf))
+* **Expeditions:** Shared elapsed time for expeditions. ([cb50be6](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/cb50be6c96cd22f0141afeee197be70d1d539ded))
+* **merger:** pesky ([ddd9aab](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/ddd9aab42632ec3177a18a065947dd4e6884cd62))
+* **prmerge:** test ([794f26d](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/794f26d4ab6b9e523ad311c9120af6a2e83acf88))
+* **status:** swapping actions stops previous action. ([1bd474e](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/1bd474ea03ba2acec9321bc6e0152fb400aa509a))
+* **versiondisplay:** updated value ([39d3e83](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/39d3e837f91c3eaa7a2ea2b9f76eada11e59e37c))
+
+## [1.7.2](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/compare/v1.7.1...v1.7.2) (2024-07-16)
+
+
+### Bug Fixes
+
+* **expeditions:** interval didn't properly stop ([83e1d14](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/83e1d149dcca4532177f5e953f708ce42867de92))
+
+## [1.7.1](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/compare/v1.7.0...v1.7.1) (2024-07-16)
+
+
+### Bug Fixes
+
+* **merger:** pesky ([ddd9aab](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/commit/ddd9aab42632ec3177a18a065947dd4e6884cd62))
+
 ## [1.7.0](https://github.com/Fishbone-Aquatics/Aquarium-Capitalists/compare/v1.6.2...v1.7.0) (2024-07-16)